fix: bugs

+++ Audentia-Players/src/main/java/fr/audentia/players/domain/model/teams/Team.java
@@ -14,9 +14,9 @@ public class Team {
     public final Map<Day, DayTransfers> transfers;
     public final Map<Day, ColiseumKills> coliseumKills;
     public final String name;
-    public final int houseId;
+    public final long houseId;
 
-    public Team(Color color, Balance balance, Map<Day, DayTransfers> transfers, Map<Day, ColiseumKills> coliseumKills, String name, int houseId) {
+    public Team(Color color, Balance balance, Map<Day, DayTransfers> transfers, Map<Day, ColiseumKills> coliseumKills, String name, long houseId) {
         this.color = color;
         this.balance = balance;
         this.transfers = transfers;
