+++ Audentia-Protect/src/main/java/fr/audentia/protect/domain/house/HouseAction.java
@@ -117,7 +117,7 @@ public class HouseAction {
             return Optional.empty();
         }
 
-        int houseId = houseRepository.getHouseId(signLocation);
+        long houseId = houseRepository.getHouseId(signLocation);
         return teamsManager.getTeamByHouseId(houseId);
     }
